Cancel running jobs when new changes arrive.
Although gomon's JobRunner supports cancellation, its invocation was
wrapped in a Once() call, such that in fact gomon never actually
cancelled running jobs. Fix this by constructing an explicit
delay runner which:

1. Supports triggering with new filenames at any time,
2. Results in job execution no later than 500ms after a new
   file event, and
3. Cancels any existing job when a new job is started.

For the common case, this seems to be the right path: even if you had a
long running build, you'd still want to restart the build given that you
had new files.

This also supports a second case of using gomon to auto-reload a server.
This use case was hinted at in the README's example

```
gomon src -- go run -x server.go
```

but only works with a change like the one in this commit -- i.e. one
where starting a new job cancels the existing, running job.